backend/requirements.txt
# Core Django
Django>=5.0,<6.0
djangorestframework>=3.15.0
djangorestframework-simplejwt>=5.2.0

# Database
psycopg2-binary>=2.9.0
dj-database-url>=2.1.0

# Security & Authentication
argon2-cffi>=23.1.0

# HTTP & API
django-cors-headers>=4.0.0
gunicorn>=21.2.0
requests>=2.31.0  # Pour webhooks (alertes critiques)

# Utilities
python-dotenv>=1.0.0
django-filter>=23.0
tenacity>=8.2.0  # Résilience intelligente pour retries avec backoff exponentiel

# Media handling
Pillow>=10.0.0
django-storages>=1.14.2
boto3>=1.34.0  # Pour S3/R2 (Cloudflare R2 compatible S3)

# Admin interface
django-jazzmin>=2.6.0

# Email
resend>=0.6.0

# Testing
pytest>=7.4.0
pytest-django>=4.7.0
pytest-cov>=4.1.0

# WebSockets & Real-time
channels>=4.0.0
channels-redis>=4.1.0
daphne>=4.0.0

# Task Queue (Asynchronous)
celery>=5.4.0
flower>=2.0.0  # Monitoring Celery (optionnel)

# Embeddings & Sémantique ⭐ NOUVEAU
openai>=1.0.0  # Optionnel: pour embeddings OpenAI
sentence-transformers>=2.2.0  # Optionnel: pour embeddings locaux

# Sécurité & Validation ⭐ NOUVEAU
pyclamd>=0.4.0  # Scan antivirus ClamAV (optionnel) - Version corrigée (max disponible: 0.4.0)
python-magic>=0.4.27  # Validation type MIME (optionnel)

# Static files
whitenoise>=6.6.0

# Security
django-csp>=3.8
cryptography>=41.0.0

# API Documentation
drf-spectacular>=0.27.0

# Security tools (dev only)
bandit>=1.9.2  # Compatible Python 3.14
safety>=2.3.5
